// src/openapi/model/api-signup-input-to-add.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiSignupInputToAdd
 */
export interface ApiSignupInputToAdd {
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    email: string;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    postalCode1: string;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    postalCode2: string;
    /**
     *
     * @type {number}
     * @memberof ApiSignupInputToAdd
     */
    areaId: number;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    address1: string;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    address2: string;
    /**
     *
     * @type {number}
     * @memberof ApiSignupInputToAdd
     */
    experienceYears: number;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    selfIntroduction: string;
    /**
     *
     * @type {string}
     * @memberof ApiSignupInputToAdd
     */
    password: string;
}